restructured Wordpress things into a module
